split worker registration from logic
